Add stb_image for image loading

+++ chaigame/graphics.h
@@ -42,8 +42,8 @@ namespace chaigame {
 		void setBackgroundColor(int red, int green, int blue, int alpha);
 		void setBackgroundColor(int red, int green, int blue);
 
-		int r = 255, g = 255, b = 255, a = 255;
-		int backR = 0, backG = 0, backB = 0, backA = 255;
+		Uint8 r = 255, g = 255, b = 255, a = 255;
+		Uint8 backR = 0, backG = 0, backB = 0, backA = 255;
 		SDL_Surface* getScreen();
 
 		int getWidth();
